security: comprehensive audit fixes and Socket.IO integration
Critical Security Fixes:
- Fix task assignment authorization: members can now unassign themselves
- Fix authentication IP extraction: read from headers not cookies (lib/auth.ts)
- Harden file upload security: prevent double-extension attacks with MIME validation
- Add JSON injection prevention: Zod schema validation in automation engine
- Implement environment validation: fail-fast on invalid configuration with Zod

Socket.IO Integration:
- Create custom server.cjs integrating Next.js with Socket.IO
- Update package.json scripts to use custom server for dev/production
- Proper authentication middleware for Socket.IO connections

New Features:
- Board export API (JSON/CSV formats)
- Board webhooks system with HMAC signature support
- Task time tracking with start/stop/manual logging
- Metrics API for board analytics

Documentation:
- Update CLAUDE.md with current architecture and security patterns
- Add environment validation system documentation
- Document custom Socket.IO server setup

This addresses 7 critical security vulnerabilities identified in comprehensive audit.

+++ lib/validations/task.ts
@@ -28,7 +28,7 @@ export const updateTaskSchema = z.object({
 
 export const moveTaskSchema = z.object({
   targetColumnId: z.string().min(1, 'Target column ID is required'),
-  newPosition: z.number().int().min(0, 'Position must be non-negative'),
+  newPosition: z.number().min(0, 'Position must be non-negative'),
   version: z.number().int().positive('Version must be positive'),
   override: z.boolean().optional(),
 })
